% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_frk.R
\docType{data}
\name{cop_frk}
\alias{cop_frk}
\alias{cop_gum}
\title{Built-In Bivariate Parametric Copula Families}
\format{An object of class \code{list} of length 7.}
\usage{
cop_frk

cop_gum
}
\description{
These objects contain information about bivariate
parametric copula families.
}
\details{
Each is a list containing the following components:
\describe{
  \item{pcop}{cdf}
  \item{dcop}{pdf}
  \item{logdcop}{log pdf}
  \item{pcondcop}{list of length 2 containing the conditional distribution
  functions of the first variable given the second, and the second variable
  given the first (respectively).}
}
If symmetries are relevant, the above applies to
direction=1, major_axis=1, and minor_axis=1 (whichever are relevant).

These objects rely heavily on functions from the package CopulaModel,
hence why coperate Depends on this package.
}
\keyword{datasets}
