% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cpar_is_oob}
\alias{cpar_is_oob}
\title{Copula parameter out of bounds}
\usage{
cpar_is_oob(cpar)
}
\arguments{
\item{cpar}{Out of bounds copula parameter. This is first
quoted to capture the copula parameter name.}
}
\value{
A string, intended to be used within the stop() function.
}
\description{
Generate a message stating that a copula
parameter is out of bounds.
}
