% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_bicop.R
\name{new_bicop}
\alias{new_bicop}
\title{Constructor Function for a Bivariate Copula}
\usage{
new_bicop(baseobj)
}
\arguments{
\item{baseobj}{List with distribution quantities.}
}
\description{
Constructor Function for a Bivariate Copula
}
