% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cop.R
\name{new_cop}
\alias{new_cop}
\title{Constructor Function for a Copula}
\usage{
new_cop(x, variable, ..., class = character())
}
\arguments{
\item{x}{List containing the components of a
copula object}

\item{variable}{Type of random variable: "continuous",
"discrete", or "mixed".}

\item{...}{Other attibutes to add to the object.}

\item{class}{If making a subclass, specify it here.}
}
\description{
Constructor Function for a Copula
}
