% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict.R
\name{restrict_symm}
\alias{restrict_symm}
\title{Restrict a copula family}
\usage{
restrict_symm(cop, ...)
}
\arguments{
\item{cop}{Object of class \code{copula}.}

\item{...}{Logical statement to perform restriction}
}
\description{
Restricts the parameter space of a copula class.
}
\details{
Names of the canonical copula parameters
}
